refactor(useRotate): 회전 로직 useRotate hook으로 분리
